+++ mono_repo/test/readme_test.dart
@@ -4,9 +4,9 @@
 
 import 'dart:io';
 
+import 'package:mono_repo/src/ci_test_script.dart';
 import 'package:mono_repo/src/commands/travis.dart';
 import 'package:mono_repo/src/package_config.dart';
-import 'package:mono_repo/src/travis_shell.dart';
 import 'package:test/test.dart';
 import 'package:test_descriptor/test_descriptor.dart' as d;
