+++ mono_repo/test/src/expected_output.dart
@@ -1,4 +1,4 @@
-import 'package:mono_repo/src/travis_shell.dart';
+import 'package:mono_repo/src/ci_test_script.dart';
 
 final travisShellOutput = '''
 #!/bin/bash
